+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Add `AdminService::getRainbowSupportBotService` `AdminService::getABotServiceData` `AdminService::getAllBotServices` to retrieve informations of bot services.
 -   Fix retrieveRainbowEntriesList
 -   Add `AdminService::createCompanyFromDefault` This API API allows to create a company for a user belonging to the 'Default' company is able to create his own company.
+-   Add `AdminService::getAllCompaniesVisibleByUser` This API allows users to get all companies.
+-   Add `AdminService::getCompanyAdministrators` This API allows users to list users being administrator of a company.
 
 ## [2.20.0] - 2023-02-24
 -   Add `PresenceService::setApplyMsTeamsPresenceSettings` This api allows to activate the exchange of presence of the connected user between rainbow and MS Teams on UI side.